# perf: add query optimizer