@mr-tz 2021-11-09
I think for samples with many functions and large functions this could make a big difference (see K32 results).